requirements.txt
transformers>=4.42.0
torch>=2.2.0
Pillow>=10.3.0
pyttsx3>=2.90
pydub>=0.25.1
tqdm>=4.66.0